fix: add pre-refresh hook to set port if unset
On a fresh install, the install hook set a default port (5000), but on a
refresh from a version without configurable port to a version with
configurable port, the configure hook would throw an error because the
configuration value for the port was unset, causing the refresh to fail.

By adding a pre-refresh hook, we ensure that there is a port set prior
to the configure hook, so the refresh can succeed.

Signed-off-by: Oliver Calder <oliver@calder.dev>