fix: move pre-refresh hook to post-refresh so the new revision sets the port
The pre-refresh hook executes on the existing revision, which
inconveniently does not have the pre-refresh hook yet set, so the
refresh fails when refreshing from a revision without support for a
configurable port. What we really need is for the new revision to have a
port set prior to the configure hook, so do so in the post-refresh hook,
which is run with the new snap revision.

Signed-off-by: Oliver Calder <oliver@calder.dev>